requirements.txt
Flask
PyJWT
bcrypt
psycopg2-binary
python-dotenv
pydantic
crewai
langchain-community
langchain-openai
duckduckgo-search
langchain
langchain-core
langchain-text-splitters
openai
pydantic
pydantic-core
requests
urllib3
uvicorn
starlette
typing-extensions
python-multipart
pyyaml
click
itsdangerous
jinja2
markupsafe
werkzeug
